requirements.txt
PyQt5
PyQt5-tools
alembic
sqlalchemy==2.0.23
mysql-connector-python
pymysql
flet==0.14.0